docs(Select): fix missing Select doc page
Move SelectTypography to seperated file
