+++ packages/vkui/src/components/Select/Select.tsx
@@ -1,43 +1,11 @@
 import * as React from 'react';
 import { NativeSelect } from '../NativeSelect/NativeSelect';
 import { CustomSelect, SelectProps } from '../CustomSelect/CustomSelect';
-import { useAdaptivity } from '../../hooks/useAdaptivity';
 import { useAdaptivityHasPointer } from '../../hooks/useAdaptivityHasPointer';
-import { usePlatform } from '../../hooks/usePlatform';
 import { classNames } from '@vkontakte/vkjs';
-import { getPlatformClassName } from '../../helpers/getPlatformClassName';
-import { getSizeYClassName } from '../../helpers/getSizeYClassName';
 import styles from './Select.module.css';
 
 export type SelectType = 'default' | 'plain' | 'accent';
-
-/**
- * @see https://vkcom.github.io/VKUI/#/SelectTypography
- */
-export const SelectTypography = ({
-  selectType = 'default',
-  children,
-  className,
-  ...restProps
-}: React.PropsWithChildren<Pick<SelectProps, 'selectType' | 'className'>>) => {
-  const platform = usePlatform();
-  const { sizeY } = useAdaptivity();
-
-  return (
-    <span
-      className={classNames(
-        styles['SelectTypography'],
-        getPlatformClassName(styles['SelectTypography'], platform),
-        getSizeYClassName(styles['SelectTypography'], sizeY),
-        styles[`SelectTypography--selectType-${selectType}`],
-        className,
-      )}
-      {...restProps}
-    >
-      {children}
-    </span>
-  );
-};
 
 /**
  * @see https://vkcom.github.io/VKUI/#/Select
