+++ packages/vkui/src/components/SelectMimicry/SelectMimicry.tsx
@@ -6,7 +6,8 @@ import { HasAlign, HasRootRef } from '../../types';
 import { usePlatform } from '../../hooks/usePlatform';
 import { getPlatformClassName } from '../../helpers/getPlatformClassName';
 import { getFormFieldModeFromSelectType } from '../../lib/select';
-import { SelectType, SelectTypography } from '../Select/Select';
+import type { SelectType } from '../Select/Select';
+import { SelectTypography } from '../SelectTypography/SelectTypography';
 import { useAdaptivity } from '../../hooks/useAdaptivity';
 import { getSizeXClassName } from '../../helpers/getSizeXClassName';
 import { getSizeYClassName } from '../../helpers/getSizeYClassName';
